refactor(grading): split semantic free-text modules
