+++ packages/grading/src/freeTextSemanticValidation.ts
@@ -1,0 +1,244 @@
+import type { FreeTextRubricSchema } from '@klicker-uzh/types'
+import {
+  approximatelyEqual,
+  hasDuplicates,
+  isFiniteNumber,
+  isNonEmptyString,
+  isRecord,
+  isStringArray,
+  SCORE_MAX,
+  SCORE_MIN,
+} from './freeTextSemanticPrimitives.js'
+
+export function validateFreeTextRubricSchema(value: unknown): string[] {
+  if (!isRecord(value)) return ['rubric schema must be an object']
+
+  const errors: string[] = []
+  if (!isNonEmptyString(value.schema_version)) {
+    errors.push('schema_version is required')
+  }
+  if (!isNonEmptyString(value.name)) errors.push('schema name is required')
+  if (!isNonEmptyString(value.description)) {
+    errors.push('schema description is required')
+  }
+  if (!Array.isArray(value.rubrics) || value.rubrics.length === 0) {
+    errors.push('at least one rubric is required')
+    return errors
+  }
+
+  const rubricIds: string[] = []
+  const rubricNames: string[] = []
+  let weightTotal = 0
+
+  value.rubrics.forEach((rubric, rubricIndex) => {
+    if (!isRecord(rubric)) {
+      errors.push(`rubric ${rubricIndex + 1} must be an object`)
+      return
+    }
+
+    if (isNonEmptyString(rubric.id)) rubricIds.push(rubric.id)
+    else errors.push(`rubric ${rubricIndex + 1} id is required`)
+
+    if (isNonEmptyString(rubric.name)) rubricNames.push(rubric.name)
+    else errors.push(`rubric ${rubricIndex + 1} name is required`)
+
+    if (!isNonEmptyString(rubric.description)) {
+      errors.push(`rubric ${rubricIndex + 1} description is required`)
+    }
+
+    if (
+      !isFiniteNumber(rubric.weight) ||
+      rubric.weight < 0 ||
+      rubric.weight > 1
+    ) {
+      errors.push(`rubric ${rubricIndex + 1} weight must be between 0 and 1`)
+    } else {
+      weightTotal += rubric.weight
+    }
+
+    if (
+      !Array.isArray(rubric.achievement_levels) ||
+      rubric.achievement_levels.length === 0
+    ) {
+      errors.push(
+        `rubric ${rubricIndex + 1} must have at least one achievement level`
+      )
+      return
+    }
+
+    const levelNames: string[] = []
+    rubric.achievement_levels.forEach((level, levelIndex) => {
+      if (!isRecord(level)) {
+        errors.push(
+          `rubric ${rubricIndex + 1} achievement level ${levelIndex + 1} must be an object`
+        )
+        return
+      }
+
+      if (isNonEmptyString(level.name)) levelNames.push(level.name)
+      else {
+        errors.push(
+          `rubric ${rubricIndex + 1} achievement level ${levelIndex + 1} name is required`
+        )
+      }
+
+      if (!isNonEmptyString(level.description)) {
+        errors.push(
+          `rubric ${rubricIndex + 1} achievement level ${levelIndex + 1} description is required`
+        )
+      }
+
+      if (
+        !isFiniteNumber(level.normalized_score) ||
+        level.normalized_score < SCORE_MIN ||
+        level.normalized_score > SCORE_MAX
+      ) {
+        errors.push(
+          `rubric ${rubricIndex + 1} achievement level ${levelIndex + 1} score must be between 0 and 100`
+        )
+      }
+    })
+
+    if (hasDuplicates(levelNames)) {
+      errors.push('achievement level names must be unique within a rubric')
+    }
+  })
+
+  if (hasDuplicates(rubricIds)) errors.push('rubric ids must be unique')
+  if (hasDuplicates(rubricNames)) errors.push('rubric names must be unique')
+  if (!approximatelyEqual(weightTotal, 1)) {
+    errors.push('rubric weights must sum to 1')
+  }
+
+  return errors
+}
+
+export function validateEvaluateFreeTextResponse({
+  value,
+  taskBundleId,
+  rubricSchema,
+}: {
+  value: unknown
+  taskBundleId: string
+  rubricSchema: unknown
+}): string[] {
+  if (!isRecord(value)) return ['evaluator response must be an object']
+  if (validateFreeTextRubricSchema(rubricSchema).length > 0) {
+    return ['rubric schema is invalid']
+  }
+
+  const schema = rubricSchema as FreeTextRubricSchema
+  const errors: string[] = []
+  if (value.contract_version !== '1') {
+    errors.push('response contract_version must be 1')
+  }
+  if (value.task_bundle_id !== taskBundleId) {
+    errors.push('response task_bundle_id does not match the request')
+  }
+  if (!isNonEmptyString(value.evaluator_version)) {
+    errors.push('response evaluator_version is required')
+  }
+  if (!isNonEmptyString(value.model_version)) {
+    errors.push('response model_version is required')
+  }
+  if (!Array.isArray(value.rubric_assessments)) {
+    errors.push('rubric assessments are required')
+    return errors
+  }
+
+  const configuredRubricById = new Map(
+    schema.rubrics.map((rubric) => [rubric.id, rubric])
+  )
+  const assessedRubricIds: string[] = []
+  value.rubric_assessments.forEach((assessment, index) => {
+    const prefix = `rubric assessment ${index + 1}`
+    if (!isRecord(assessment)) {
+      errors.push(`${prefix} must be an object`)
+      return
+    }
+
+    if (assessment.task_bundle_id !== taskBundleId) {
+      errors.push(`${prefix} task_bundle_id does not match the request`)
+    }
+
+    const rubricId = assessment.rubric_id
+    if (!isNonEmptyString(rubricId)) {
+      errors.push(`${prefix} rubric_id is required`)
+    } else {
+      assessedRubricIds.push(rubricId)
+      const configuredRubric = configuredRubricById.get(rubricId)
+      if (!configuredRubric) {
+        errors.push(`${prefix} rubric_id is not configured`)
+      } else {
+        if (assessment.rubric_name !== configuredRubric.name) {
+          errors.push(`${prefix} rubric_name does not match the rubric`)
+        }
+        if (
+          !isNonEmptyString(assessment.proposed_level) ||
+          !configuredRubric.achievement_levels.some(
+            (level) => level.name === assessment.proposed_level
+          )
+        ) {
+          errors.push(`${prefix} proposed_level is not configured`)
+        }
+      }
+    }
+
+    if (
+      !isFiniteNumber(assessment.normalized_score) ||
+      assessment.normalized_score < SCORE_MIN ||
+      assessment.normalized_score > SCORE_MAX
+    ) {
+      errors.push(`${prefix} score must be between 0 and 100`)
+    }
+    if (
+      !isFiniteNumber(assessment.confidence) ||
+      assessment.confidence < 0 ||
+      assessment.confidence > 1
+    ) {
+      errors.push(`${prefix} confidence must be between 0 and 1`)
+    }
+    if (assessment.needs_review === true) {
+      errors.push(`${prefix} requires human review`)
+    } else if (assessment.needs_review !== false) {
+      errors.push(`${prefix} needs_review must be false`)
+    }
+
+    for (const [field, fieldValue] of [
+      ['evidence_ids', assessment.evidence_ids],
+      ['review_flags', assessment.review_flags],
+      ['used_evidence_ids', assessment.used_evidence_ids],
+      ['unsupported_claims', assessment.unsupported_claims],
+    ] as const) {
+      if (!isStringArray(fieldValue)) {
+        errors.push(`${prefix} ${field} must be a string array`)
+      }
+    }
+    if (!isNonEmptyString(assessment.justification)) {
+      errors.push(`${prefix} justification is required`)
+    }
+    if (!isNonEmptyString(assessment.rationale)) {
+      errors.push(`${prefix} rationale is required`)
+    }
+  })
+
+  const containsEveryRubricExactlyOnce =
+    assessedRubricIds.length === schema.rubrics.length &&
+    schema.rubrics.every((rubric) => {
+      return assessedRubricIds.filter((id) => id === rubric.id).length === 1
+    })
+  if (!containsEveryRubricExactlyOnce) {
+    errors.push(
+      'rubric assessments must contain every configured rubric exactly once'
+    )
+  }
+
+  if (
+    value.feedback_proposals !== undefined &&
+    !Array.isArray(value.feedback_proposals)
+  ) {
+    errors.push('feedback proposals must be an array')
+  }
+
+  return errors
+}
